pyproject.toml
[project]
name = "rest-food"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "channels[daphne]>=4.2.0",
    "djangorestframework>=3.15.2",
    "python-decouple>=3.8",
]
